extracted StateVector and CircuitMatrix to new module for cleaner structure and possibility to use their stringification when showing gate-matrices in the Help menu
