+++ qrogue/game/logic/__init__.py
@@ -1,7 +1,7 @@
 # exporting
 from qrogue.game.logic.message import Message
-from qrogue.game.logic.actors import StateVector
 
+import qrogue.game.logic.base
 import qrogue.game.logic.collectibles
 import qrogue.game.logic.actors
