+++ qrogue/game/logic/actors/puzzles/enemy.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from qrogue.game.logic.actors import StateVector
+from qrogue.game.logic.base import StateVector
 from qrogue.game.logic.collectibles import Collectible
 from qrogue.util import RandomManager, PuzzleConfig, MyRandom
