+++ qrogue/game/logic/base/__init__.py
@@ -1,0 +1,2 @@
+from .state_vector import StateVector
+from .circuit_matrix import CircuitMatrix
