+++ qrogue/graphics/widgets/my_widgets.py
@@ -6,8 +6,8 @@ from py_cui import ColorRule
 from py_cui.widget_set import WidgetSet
 from py_cui.widgets import BlockLabel
 
-from qrogue.game.logic import StateVector
-from qrogue.game.logic.actors import Robot, CircuitMatrix
+from qrogue.game.logic.base import StateVector, CircuitMatrix
+from qrogue.game.logic.actors import Robot
 from qrogue.game.logic.collectibles import Instruction
 from qrogue.game.world.map import Map
 from qrogue.game.world.navigation import Direction
