+++ qrogue/test/statevector_tests.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from qrogue.game.logic.actors import StateVector
+from qrogue.game.logic.base import StateVector
 
 stv = StateVector([1 / np.sqrt(2), 0 + 0j, 0 + 0j, 1 / np.sqrt(2)])
 #stv.extend(1)
